ai-crypto-analyzer/requirements.txt
requests==2.31.0
google-generativeai==0.7.1
python-telegram-bot==20.7
